+++ src/BehaviorTree/BlackBoard.cpp
@@ -11,13 +11,13 @@ CBlackBoard::~CBlackBoard()
 }
 
 //=======================================================================================================================
-void CBlackBoard::SetValue(std::string InKey, CBlackBoardValueBase* InValue, int InTreeId, int InTaskId)
+void CBlackBoard::SetValue(std::string InKey, CBlackBoardValueBase* InValue, size_t InTreeId, size_t InTaskId)
 {
 	NodeMap[InTreeId][InTaskId][InKey].reset(InValue);
 }
 
 //=======================================================================================================================
-void CBlackBoard::SetValue(std::string InKey, CBlackBoardValueBase* InValue, int InTreeId)
+void CBlackBoard::SetValue(std::string InKey, CBlackBoardValueBase* InValue, size_t InTreeId)
 {
 	TreeMap[InTreeId][InKey].reset(InValue);
 }
@@ -29,13 +29,13 @@ void CBlackBoard::SetValue(std::string InKey, CBlackBoardValueBase* InValue)
 }
 
 //=======================================================================================================================
-std::shared_ptr<CBlackBoardValueBase> CBlackBoard::GetValue(std::string InKey, int InTreeId, int InTaskId)
+std::shared_ptr<CBlackBoardValueBase> CBlackBoard::GetValue(std::string InKey, size_t InTreeId, size_t InTaskId)
 {
 	return NodeMap[InTreeId][InTaskId][InKey];
 }
 
 //=======================================================================================================================
-std::shared_ptr<CBlackBoardValueBase> CBlackBoard::GetValue(std::string InKey, int InTreeId)
+std::shared_ptr<CBlackBoardValueBase> CBlackBoard::GetValue(std::string InKey, size_t InTreeId)
 {
 	return TreeMap[InTreeId][InKey];
 }
